requirements.txt
pandas
pre-commit
pytorch-lightning
torchaudio
torchvision
wandb
miniaudio